created hashed qr code ande removed register component
