+++ backend/package.json
@@ -15,8 +15,8 @@
         "@babel/preset-env": "^7.23.9",
         "@babel/preset-react": "^7.23.3",
         "@babel/register": "^7.23.7",
+        "bcrypt": "^5.1.1",
         "cors": "^2.8.5",
-        "crypto": "^1.0.1",
         "dotenv": "^16.3.2",
         "express": "^4.18.2",
         "express-session": "^1.18.0",
