+++ frontend/package.json
@@ -26,6 +26,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-otp-input": "^3.1.1",
+    "react-qr-code": "^2.0.12",
     "react-qr-reader": "^3.0.0-beta-1",
     "react-router-dom": "^6.21.2",
     "react-stickynode": "^4.1.1",
